Addressed PR comments including sending 'rows affected' messages at the correct time and sending a 'success' message if no rows affected and an error did not occur
